fixing warnings refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@15115 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSVehicle.cpp
@@ -1707,10 +1707,10 @@ MSVehicle::getBestLanes(bool forceRebuild, MSLane* startLane) const {
                 // patch invalid bestLaneOffset and updated myCurrentLaneInBestLanes
                 for (size_t i = 0; i < lanes.size(); ++i) {
                     if (i + lanes[i].bestLaneOffset < 0) {
-                        lanes[i].bestLaneOffset = -i;
+                        lanes[i].bestLaneOffset = (int)-i;
                     }
                     if (i + lanes[i].bestLaneOffset >= lanes.size()) {
-                        lanes[i].bestLaneOffset = lanes.size() - i - 1;
+                        lanes[i].bestLaneOffset = (int)(lanes.size() - i - 1);
                     }
                     assert(i + lanes[i].bestLaneOffset >= 0);
                     assert(i + lanes[i].bestLaneOffset < lanes.size());
